+++ CHANGELOG.md
@@ -1,19 +1,34 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2020-09-08
+
+### Added
+
+- Schema character length diff compared to previous version (red/green) in schema version list
+
+### Fixed
+
+- Increased max schema type_defs in body_parser middleware and in DB from ~100KB to ~16MB
+
 ## [1.0.1] - 2020-09-03
+
 ### Security
+
 - Updated prismjs version 1.19.0 to 1.21.0
 
 ## [1.0.0] - 2020-09-03
+
 ### Added
+
 - Backend service
 - Frontend app
 - Examples of gateway + 2 federated services
 
-[Unreleased]: https://github.com/pipedrive/graphql-schema-registry/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/pipedrive/graphql-schema-registry/compare/v1.0.1...HEAD
 [1.0.1]: https://github.com/pipedrive/graphql-schema-registry/compare/v1.0.1...v1.0.1
 [1.0.0]: https://github.com/pipedrive/graphql-schema-registry/compare/v1.0.0...v1.0.0